Refactor component imports and restructure components for improved organization
