+++ src/components/FAQCard/index.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore"
 import {
   Accordion,
