+++ src/components/Links/LinkBehavior/index.tsx
@@ -1,4 +1,3 @@
-"use client"
 import { default as NextLink, LinkProps as NextLinkProps } from "next/link"
 import { forwardRef } from "react"
