--- src/server/components/signIn.tsx
@@ -1,14 +1,0 @@
-import { signIn } from "@/auth"
-
-export default function SignIn() {
-  return (
-    <form
-      action={async () => {
-        "use server"
-        await signIn("discord")
-      }}
-    >
-      <button type="submit">Signin with Discord</button>
-    </form>
-  )
-}
